mod_rate_limit: pull real client IP from request (#55)
The session ip field, especially before the connection is fully initialized (e.g. at the point that filter hooks get called) is not always accurate. For example, it may be set to the network IP that the connection came from, rather than the IP of the remote client (which may be determined by HTTP headers).

Prosody 0.12.3 will provide an API to access the request object, allowing access to the client's IP. For older versions, I've included the snippet of code in that helper function directly in this module. As it uses internal "private" implementation details, it should be removed if/when compatibility with older 0.12.x versions is not needed.