fix(visitors): Fixes meeting id in vnode events.
First jicofo connects and we generate a new meeting_id, later we receive the connect message from the main prosody to setup the room and prepare it for visitors and it is when we receive the meeting_id of the main room.
Make sure we update it in already generated conference details.